step_85=FIX_SQL

+++ student_snezhko_leu/src/main/java/lv/javaguru/travel/insurance/core/domain/MedicalRiskLimitLevel.java
@@ -6,6 +6,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -23,5 +24,5 @@ public class MedicalRiskLimitLevel {
     private String medicalRiskLimitLevelIc;
 
     @Column(name = "COEFFICIENT", nullable = false)
-    private Double coefficient;
+    private BigDecimal coefficient;
 }
